Type changes

+++ src/strutil_indexesOf.c
@@ -17,13 +17,13 @@
  *   Free the returned array when done.
  * Returns: an array of all the indexes that 'character' is found at.
  */
-int* strutil_indexesOf(char* string, char character, size_t* amount) {
+size_t* strutil_indexesOf(char* string, char character, size_t* amount) {
 	#ifdef STRUTIL_DEBUG
 	printf(COLOR_YELLOW "STRUTIL: INDEXES_OF: Searching \"%s\" for all occurences of \'%c\'\n" COLOR_RESET, string, character);
 	#endif
 
 	size_t length = strlen(string);
-	int buff[length];
+	size_t buff[length];
 	memset(buff, 0, length);
 
 	for (size_t i = 0; i < length; i++) {
@@ -36,7 +36,7 @@ int* strutil_indexesOf(char* string, char character, size_t* amount) {
 		}
 	}
 
-	int* indexes = calloc(*amount, sizeof(int));
+	size_t* indexes = calloc(*amount, sizeof(size_t));
 	for (size_t i = 0; i < (*amount); i++) {
 		indexes[i] = buff[i];
 	}
